% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateWangSimilarity.R
\name{CalculateWangSimilarity}
\alias{CalculateWangSimilarity}
\title{Calculate pairwise Wang semantic similarity between a group of genes}
\usage{
CalculateWangSimilarity(
  OrgDb,
  keytype = "ENTREZID",
  ont = "BP",
  genes = NULL,
  dropCodes = NULL,
  minIC = NULL
)
}
\arguments{
\item{OrgDb}{OrgDb object}

\item{keytype}{keytype to select records from OrgDb. All possible keys are returned by using the keys method.}

\item{ont}{Type of ontology to be used. One of 'BP', 'MF', 'CC'. Defaults to BP.}

\item{genes}{List of genes for which to calculate pairwise Wang semantic similarity between them. Defaults to NULL (use all annotated genes)}

\item{dropCodes}{List of codes of annotation evidence types. Annotations with this codes will be dropped before calculating similiarity. Defaults to NULL, including all annotations.}

\item{minIC}{Minimum term information content 'IC' to take into account. Terms with less IC won't be used for calculations. Defaults to NULL}
}
\value{
A matrix of pairwise gene similarities
}
\description{
Calculate pairwise Wang semantic similarity between a group of genes as described in
James Z. Wang, Zhidian Du, Rapeeporn Payattakool, Philip S. Yu, Chin-Fu Chen,
A new method to measure the semantic similarity of GO terms,
Bioinformatics, Volume 23, Issue 10, 15 May 2007, Pages 1274–1281,
https://doi.org/10.1093/bioinformatics/btm087
}
\keyword{Wang}
\keyword{similarity}
